Refactor Code (deletion)
